fix(cms): change to json field
